fix: moving tests into test_eda NOT test_custom_cleaning
