+++ tests/data_audit/test_eda.py
@@ -1,9 +1,12 @@
 from pathlib import Path
+import logging
+import types
 
 import pandas as pd
 import pytest
 
 from edvise import data_audit, dataio
+from edvise.data_audit import eda
 from edvise.data_audit.eda import EdaSummary, log_grade_distribution
 
 
@@ -812,3 +815,187 @@ def test_compute_gateway_course_ids_prefix_plus_digits_after_letters():
     )
     assert has_upper is False
     assert "ENG10000" in lower_ids
+
+
+# -------------------------------------------------------------------
+# print_column_pct_breakdowns
+# -------------------------------------------------------------------
+def test_print_column_pct_breakdowns_logs_mapped_labels(caplog):
+    """Test that print_column_pct_breakdowns logs labeled percent breakdowns."""
+    caplog.set_level(logging.INFO, logger="edvise.data_audit.eda")
+
+    df = pd.DataFrame(
+        {
+            "enrollment_type": ["FIRST-TIME", "FIRST-TIME", "TRANSFER-IN"],
+            "intended_program_type": ["Associate", "Certificate", "Associate"],
+        }
+    )
+
+    eda.print_column_pct_breakdowns(df, eda.ES_COHORT_PCT_BREAKDOWN_COLS)
+
+    log_message = " ".join(r.getMessage() for r in caplog.records)
+    assert "Percent breakdown for enrollment types" in log_message
+    assert "Percent breakdown for intended program types" in log_message
+    assert "FIRST-TIME" in log_message
+    assert "Associate" in log_message
+
+
+def test_print_column_pct_breakdowns_skips_missing_columns(caplog):
+    """Test that missing columns are skipped with a warning."""
+    caplog.set_level(logging.WARNING, logger="edvise.data_audit.eda")
+
+    df = pd.DataFrame({"enrollment_type": ["FIRST-TIME", "TRANSFER-IN"]})
+
+    eda.print_column_pct_breakdowns(df, eda.ES_COHORT_PCT_BREAKDOWN_COLS)
+
+    messages = [r.getMessage() for r in caplog.records]
+    assert any("intended_program_type" in msg for msg in messages)
+
+
+# -------------------------------------------------------------------
+# warn_null_graduation_target_columns
+# -------------------------------------------------------------------
+def test_warn_null_graduation_target_columns_noop_for_non_graduation(caplog):
+    caplog.set_level(logging.WARNING, logger="edvise.data_audit.eda")
+    df = pd.DataFrame({"first_year_to_associates_at_cohort_inst": [None, None]})
+    target = types.SimpleNamespace(
+        type_="retention",
+        years_to_degree_col="first_year_to_associates_at_cohort_inst",
+    )
+
+    eda.warn_null_graduation_target_columns(df, target)
+
+    assert not any("Graduation model" in r.getMessage() for r in caplog.records)
+
+
+def test_warn_null_graduation_target_columns_warns_when_all_null(caplog):
+    caplog.set_level(logging.WARNING, logger="edvise.data_audit.eda")
+    df = pd.DataFrame({"first_year_to_associates_at_cohort_inst": [None, None]})
+    target = types.SimpleNamespace(
+        type_="graduation",
+        years_to_degree_col="first_year_to_associates_at_cohort_inst",
+    )
+
+    eda.warn_null_graduation_target_columns(df, target)
+
+    assert any(
+        "first_year_to_associates_at_cohort_inst" in r.getMessage()
+        and "100% null" in r.getMessage()
+        for r in caplog.records
+    )
+
+
+def test_warn_null_graduation_target_columns_warns_when_missing(caplog):
+    caplog.set_level(logging.WARNING, logger="edvise.data_audit.eda")
+    df = pd.DataFrame({"other_col": [1, 2]})
+    target = types.SimpleNamespace(
+        type_="graduation",
+        years_to_degree_col="first_year_to_associates_at_cohort_inst",
+    )
+
+    eda.warn_null_graduation_target_columns(df, target)
+
+    assert any("missing from cohort data" in r.getMessage() for r in caplog.records)
+
+
+def test_warn_null_graduation_target_columns_no_warning_when_populated(caplog):
+    caplog.set_level(logging.WARNING, logger="edvise.data_audit.eda")
+    df = pd.DataFrame({"first_year_to_associates_at_cohort_inst": [1, 2, None]})
+    target = types.SimpleNamespace(
+        type_="graduation",
+        years_to_degree_col="first_year_to_associates_at_cohort_inst",
+    )
+
+    eda.warn_null_graduation_target_columns(df, target)
+
+    assert not any("Graduation model" in r.getMessage() for r in caplog.records)
+
+
+# -------------------------------------------------------------------
+# log_top_majors
+# -------------------------------------------------------------------
+def test_log_top_majors_logs_top_10(caplog):
+    """Test that log_top_majors logs the top 10 majors."""
+    caplog.set_level(logging.INFO, logger="edvise.data_audit.eda")
+
+    df = pd.DataFrame(
+        {
+            "program_of_study_term_1": ["CS"] * 50
+            + ["EE"] * 30
+            + ["ME"] * 20
+            + ["CE"] * 15
+            + ["BIO"] * 10
+            + ["CHEM"] * 8
+            + ["PHYS"] * 6
+            + ["MATH"] * 4
+            + ["STAT"] * 3
+            + ["DATA"] * 2
+            + ["ART"] * 1,
+        }
+    )
+
+    eda.log_top_majors(df)
+
+    messages = [r.getMessage() for r in caplog.records]
+    log_message = " ".join(messages)
+
+    assert "CS" in log_message
+    assert "EE" in log_message
+    assert "Top majors" in log_message
+
+
+def test_log_top_majors_handles_less_than_10_majors(caplog):
+    """Test that log_top_majors handles dataframes with less than 10 majors."""
+    caplog.set_level(logging.INFO, logger="edvise.data_audit.eda")
+
+    df = pd.DataFrame(
+        {
+            "program_of_study_term_1": ["CS", "CS", "EE", "EE", "ME"],
+        }
+    )
+
+    eda.log_top_majors(df)
+
+    messages = [r.getMessage() for r in caplog.records]
+    assert any("Top majors" in msg for msg in messages)
+
+
+def test_log_top_majors_handles_null_values(caplog):
+    """Test that log_top_majors includes null values in counts."""
+    caplog.set_level(logging.INFO, logger="edvise.data_audit.eda")
+
+    df = pd.DataFrame(
+        {
+            "program_of_study_term_1": ["CS"] * 10 + [None] * 5 + ["EE"] * 3,
+        }
+    )
+
+    eda.log_top_majors(df)
+
+    messages = [r.getMessage() for r in caplog.records]
+    assert any("Top majors" in msg for msg in messages)
+
+
+def test_log_top_majors_custom_column(caplog):
+    """Test that log_top_majors supports a schema-specific major column."""
+    caplog.set_level(logging.INFO, logger="edvise.data_audit.eda")
+
+    df = pd.DataFrame({"declared_major_at_entry": ["Nursing", "Nursing", "Business"]})
+
+    eda.log_top_majors(df, major_col="declared_major_at_entry")
+
+    log_message = " ".join(r.getMessage() for r in caplog.records)
+    assert "declared_major_at_entry" in log_message
+    assert "Nursing" in log_message
+
+
+def test_log_top_majors_returns_none():
+    """Test that log_top_majors returns None (it's a logging function)."""
+    df = pd.DataFrame(
+        {
+            "program_of_study_term_1": ["CS", "EE"],
+        }
+    )
+
+    result = eda.log_top_majors(df)
+    assert result is None
